webroot content merged with client-server handler.
